Fix onnxruntime output access (List, not Map) and Offset double casts
OrtSession.run()/runAsync() returns List<OrtValue?>, not a Map --
fingertip_onnx_service.dart and classifier_service.dart were indexing
it with String keys ('output0'/'output') and calling .values, both of
which are undefined on a List and broke the release build. Fixed to
outputs[0]?.value / iterate the list directly for release; removed
classifier_service.dart's now-dead defensive 'outputs is Map' branch.

learning_screen.dart: num-typed ix/iy (from .clamp(0, ...)) didn't
satisfy Offset's strict double parameters -- added .toDouble().

Also cleared the remaining flutter analyze lints (unused imports/
variable, a private type in a public API) so analyze is fully clean.

+++ braille_lens_flutter/lib/screens/learning_screen.dart
@@ -576,7 +576,7 @@ class _TapFingertipDialogState extends State<_TapFingertipDialog> {
                   final oy = (360 - dh) / 2;
                   final ix = ((local.dx - ox) / s).clamp(0, _image!.width.toDouble());
                   final iy = ((local.dy - oy) / s).clamp(0, _image!.height.toDouble());
-                  Navigator.pop(context, Offset(ix, iy));
+                  Navigator.pop(context, Offset(ix.toDouble(), iy.toDouble()));
                 },
                 child: CustomPaint(
                   painter: _ImagePainter(_image!),
